requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
pytesseract
python-multipart
pillow